requirements.txt
streamlit==1.22.0
streamlit-option-menu==0.3.6
pandas==2.0.0
plotly==5.14.1
sqlite3==3.41.2
geopy==2.3.0
Pillow==9.5.0
inference-sdk==0.0.10